Fallback to CPU for mod only for DB11.3
